Setup files for testing
