+++ assets/state.js
@@ -1,0 +1,13 @@
+const rootUrl = "https://unfiltered-app.herokuapp.com/";
+// const rootUrl = 'http://localhost:3000/'
+// set state of the app
+const state = {
+    data: [
+    ],
+    comments: [
+    ]
+}
+
+/**
+ * NO TESTING NEEDED HERE
+ */
