+++ index.html
@@ -9,6 +9,11 @@
 
     <script src="https://kit.fontawesome.com/67eb281458.js" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.7.7/handlebars.min.js" defer></script>
+    <script src="./assets/state.js" defer></script>
+    <script src="./assets/listeners.js" defer></script>
+    <script src="./assets/handlebars.js" defer></script>
+    <script src="./assets/reactions.js" defer></script>
+    <script src="./assets/newEntry.js" defer></script>
     <script src="./assets/script.js" defer></script>
