+++ package.json
@@ -1,5 +1,11 @@
 {
   "dependencies": {
     "handlebars": "^4.7.7"
+  },
+  "scripts": {
+    "test": "jest --watchAll --coverage"
+  },
+  "devDependencies": {
+    "jest": "^28.1.3"
   }
 }
